% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicParam-class.R
\name{deconvolution,epicParam-method}
\alias{deconvolution,epicParam-method}
\title{Deconvolution method for class \linkS4class{epicParam}}
\usage{
\S4method{deconvolution}{epicParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{epicParam}.}
}
\value{
Either a vector of predicted proportions, or a list containing 
predictions, metadata, and original outputs.
}
\description{
Main deconvolution method for the \linkS4class{epicParam} to run the 
\code{EPIC::EPIC()} implementation of the EPIC algorithm.
}
\examples{
exampleDataList <- getDeconvolutionExampleData()
newParam <- epicParam(exampleDataList[["bulkExpression"]],exampleDataList[["referenceExpression"]],exampleDataList[["cellScaleFactors"]])
deconvolution(newParam)

}
\references{
Racle, Julien, and David Gfeller. 2020. “EPIC: A Tool to Estimate the 
Proportions of Different Cell Types from Bulk Gene Expression Data.” In 
Bioinformatics for Cancer Immunotherapy: Methods and Protocols, edited by 
Sebastian Boegel, 233–48. Methods in Molecular Biology. New York, NY: 
Springer US. https://doi.org/10.1007/978-1-0716-0327-7_17.
}
