% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicParam-class.R
\docType{class}
\name{epicParam-class}
\alias{epicParam-class}
\title{epicParam-class}
\description{
Applies the EPIC::EPIC() deconvolution algorithm.
}
\details{
Main constructor for class \linkS4class{epicParam}.
}
\examples{
exampleDataList <- getDeconvolutionExampleData()
newParam <- epicParam(exampleDataList[["bulkExpression"]],exampleDataList[["referenceExpression"]],exampleDataList[["cellScaleFactors"]])
deconvolution(newParam)

}
\references{
Racle, Julien, and David Gfeller. 2020. “EPIC: A Tool to Estimate the 
Proportions of Different Cell Types from Bulk Gene Expression Data.” In 
Bioinformatics for Cancer Immunotherapy: Methods and Protocols, edited by 
Sebastian Boegel, 233–48. Methods in Molecular Biology. New York, NY: 
Springer US. https://doi.org/10.1007/978-1-0716-0327-7_17.
}
\seealso{
\linkS4class{referencebasedParam-class}
}
